Connect GPMapreduce to the build
This adds a new switch to autoconf, --enable-mapreduce, to connect
the GPMapreduce extension to the build. If set to on the mapreduce
extension in gpAux/extensions will be built and installed.  As a
side effect this will properly clean and distclean for gpfdist in
gpAux/extensions as well. Autoconf checks for libyaml since it's
a depedency for GPMapreduce.