Resolve FIXME regarding buffer reuse
Judging by the previous code which had a window for deadlock, for
now it's a better choice to not try and be clever with buffer reuse.
Since the comment has a valuable idea I changed the FIXME to a TODO
instead.

Also added back the guard for MIRROREDLOCK_BUFMGR which is still
required but was lost in the merge.