Fix EXPLAIN ANALYZE support for tuplesort_mk.
New tuplesort_explain() function was added to tuplesort.c in the merge.
Need to add it for tuplesort_mk.c, too, and teach explain.c to call
the MK variant if the gp_enable_mk_sort GUC is set. Otherwise, we pass
TupleStoreState_mk to the regular tuplesort_explain() function, which
will crash.