Resolve FIXME in cdbmutate
Passing a NULL econtext is safe in this call since getCurrentOf()
only need it to extract the cursor_name in case it's not in the
CurrentOf node (which it is in this case). Add a null pointer
deref guard to getCurrentOf() and a cautionary comment to reduce
the risk of segfaults from copy/paste.