Disable write support for old-format AO table segments.
Segments that are still in an old format are treated as read-only. All new
data go to new segments, in new format. This allows us to eventually get
rid of the old format completely.

This is hypothetical until we have pg_upgrade working for GPDB 4.3 -> 5.0
upgrade, as you can't have old-format tables or segments at all in a
cluster that's initialized with 5.0. Stay tuned for pg_upgrade, but this
is preparatory work for that.