Create new OID in QD and pass new OIDS in dump for pg_upgrade on QE
When upgrading from GPDB4 to 5, we need to create new arraytypes
for the base relation rowtypes in the QD, but we also need to
dispatch these new OIDs to the QEs. Allow for array types to carry
InvalidOid in the oid_dispatcher and instruct the backend to get a
new Oid if in binary upgrade mode and an InvalidOid is encountered.
In pg_upgrade, finish the process by dumping the new Oids into a
separate dumpfile which isn't unlinked on exit. If this file is
placed into the cwd of pg_upgrade on the QEs, it will be pulled
into the db dump and used during restoring thus "dispatching" the
Oids from the QD.

pg_upgrade doesn't at this point know if it's running at a QD or
a QE so it will always dump this file and include the InvalidOid
markers, this is a fixme that needs to be handled before merging
to master.