Grab an unused pid by looking back in time
Rather than looking ahead to get the next unused pid, look backwards
and grab a pid from a process which is guaranteed to have exited by
the time of reading. The command does a nonsensical backgrounding of
a process which exits immediatelt and takes the pid from that.

Looking ahead introduce the risk of a time-of-check-time-of-use race
since the pid might have been allocated by the operating system by
the time the test use the data.