Support per-segment tablespace location definitions
The old filespace concept allowed for different physical location
per segment/mirror, whereas the tablespace definition only support
a single location.  Tablespaces are backed by a symlink which can
be changed without any changes required on the database level, as
long as the system is turned off at the time. To recreate the
filespace functionality, a process like this would be required:

  1. CREATE TABLESPACE foo LOCATION '/foo/bar';
  2. Shut down the cluster
  3. Move the symlinks to preferred location on each segment
  4. Start cluster back up

Since step 3 require accessing each segment in the cluster, it is
preferrable to allow syntax do to this automatically. This commit
adds support for specifying per segment overrides from the default
LOCATION using an optional OPTIONS clause:

  CREATE TABLESPACE foo LOCATION '/foo/bar'
    OPTIONS (segment1 '/segment1/bar', segment2 '/segment2/bar');

Any segment not listed in the OPTIONS clause will get the LOCATION
as its tablspace symlink. The spclocation will be set to LOCATION,
but that column will shortly go away as it's no longer used.