Support equijoin on coerced distribution keys
Distribution key joins on pathkeys where the parser invokes an implicit
cast, or where the user invokes an explicit cast, would introduce a
redistribution motion due to the join not being recognized as a colocated
equijoin.

As an example, consider the following relation

    create table test (a varchar(10), b varchar(10)) distributed by (a);

The below query will introduce an implicit cast on a.a and b.a to TEXT by
the parser. This means that the pathkey EquivalenceClass is on VARCHAR but
the RestrictInfo ECs are for TEXT.

    select * from test a, test b where a.a=b.a;

Similarly, an explicit cast would also introduce a redistribution motion
as the ECs didn't match.

    select * from test a, test b where a.a::numeric=b.a::numeric;

Fix by following the EC members and checking for coercion conditions under
which we can allow colocated equijoins.